requirements.txt
tqdm
torch
torchvision
scikit-learn
PyYAML
xarray
netcdf4
zarr
dask

nvidia-physicsnemo
